Moved the logic to wal_uploader.go
